requirements.txt
absl-py==0.9.0
asgiref==3.2.5
astor==0.8.1
astroid==2.3.3
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==3.0.4
django-crispy-forms==1.9.0
djangorestframework==3.11.0
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.0
h5py==2.10.0
idna==2.8
isort==4.3.21
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pylint==2.4.4
pypiwin32==223
pyttsx3==2.71
pytz==2019.3
pywin32==227
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
sqlparse==0.3.1
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tflearn==0.3.2
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==0.16.1
wikipedia==1.4.0
wrapt==1.11.2
